refact(components): organized folder structure
