+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import App from './App'
+import App from './components/Calendar'
 
 const rootElement = document.getElementById('root')
 ReactDOM.render(
